Split new model extensions by separate files
